requirements.txt
fastapi
uvicorn
jinja2
aiofiles
python-multipart
opencv-python
ultralyticsgdown
gdown